requirements.txt
# FP&A Reporting Automation — Python Dependencies
# Install: pip install -r requirements.txt

pandas>=2.0
numpy>=1.24
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
openpyxl>=3.1          # For Excel export if needed
